requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
requests==2.31.0
aiohttp==3.9.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.26.0
pytest==7.4.3
pytest-asyncio==0.23.3
anthropic==0.18.1
openai==1.10.0
google-generativeai==0.3.2
